Speculatively work around issue with linking concurrency libs (#4323)
It looks like `swift-build` started crashing in the nightly toolchains since we introduced concurrency into SwiftPM:

```
objc[47001]: Class _TtCs27_KeyedEncodingContainerBase is implemented in both /usr/lib/swift/libswiftCore.dylib (0x7ff851b7d558) and /Library/Developer/Toolchains/swift-DEVELOPMENT-SNAPSHOT-2022-04-23-a.xctoolchain/usr/lib/swift/macosx/libswiftCore.dylib (0x10467ed38). One of the two will be used. Which one is undefined.```

Revert "update example to use async await APIs (#4263)"

This reverts commit 745ce767f82c6d4846f2b003533e8502f953939d.
