+++ Examples/package-info/Package.swift
@@ -1,25 +1,22 @@
-// swift-tools-version:5.5
+// swift-tools-version:5.1
 
 import PackageDescription
 
 let package = Package(
     name: "package-info",
     platforms: [
-        .macOS(.v12),
+        .macOS(.v10_15),
         .iOS(.v13)
     ],
     dependencies: [
         // This just points to the SwiftPM at the root of this repository.
-        .package(name: "swift-package-manager", path: "../../"),
+        .package(path: "../../"),
         // You will want to depend on a stable semantic version instead:
         // .package(url: "https://github.com/apple/swift-package-manager", .exact("0.4.0"))
     ],
     targets: [
-        .executableTarget(
+        .target(
             name: "package-info",
-            dependencies: [
-                .product(name: "SwiftPM", package: "swift-package-manager")
-            ]
-        ),
+            dependencies: ["SwiftPM"]),
     ]
 )
