+++ Sources/Workspace/Workspace.swift
@@ -1156,16 +1156,6 @@ extension Workspace {
     }
 
     /// Loads and returns manifests at the given paths.
-    @available(macOS 12, iOS 15, tvOS 15, watchOS 8, *)
-    public func loadRootManifests(packages: [AbsolutePath], observabilityScope: ObservabilityScope) async throws -> [AbsolutePath: Manifest] {
-        return try await withCheckedThrowingContinuation{ continuation in
-            self.loadRootManifests(packages: packages, observabilityScope: observabilityScope) { result in
-                continuation.resume(with: result)
-            }
-        }
-    }
-
-    /// Loads and returns manifests at the given paths.
     public func loadRootManifests(
         packages: [AbsolutePath],
         observabilityScope: ObservabilityScope,
@@ -1207,20 +1197,10 @@ extension Workspace {
     }
 
     /// Loads and returns manifest at the given path.
-    @available(macOS 12, iOS 15, tvOS 15, watchOS 8, *)
-    public func loadRootManifest(at path: AbsolutePath, observabilityScope: ObservabilityScope) async throws -> Manifest {
-        return try await withCheckedThrowingContinuation{ continuation in
-            self.loadRootManifest(at: path, observabilityScope: observabilityScope) { result in
-                continuation.resume(with: result)
-            }
-        }
-    }
-
-    /// Loads and returns manifest at the given path.
     public func loadRootManifest(
         at path: AbsolutePath,
         observabilityScope: ObservabilityScope,
-        completion: @escaping (Result<Manifest, Error>) -> Void
+        completion: @escaping(Result<Manifest, Error>) -> Void
     ) {
         self.loadRootManifests(packages: [path], observabilityScope: observabilityScope) { result in
             completion(result.tryMap{
@@ -1256,17 +1236,6 @@ extension Workspace {
         )
     }
 
-    /// Loads root package
-    @available(macOS 12, iOS 15, tvOS 15, watchOS 8, *)
-    public func loadRootPackage(at path: AbsolutePath, observabilityScope: ObservabilityScope) async throws -> Package {
-        return try await withCheckedThrowingContinuation{ continuation in
-            self.loadRootPackage(at: path, observabilityScope: observabilityScope) { result in
-                continuation.resume(with: result)
-            }
-        }
-    }
-
-    /// Loads root package
     public func loadRootPackage(
         at path: AbsolutePath,
         observabilityScope: ObservabilityScope,
